+++ tests/operators/where.test.ts
@@ -2,6 +2,7 @@ import { describe, it, expect } from 'vitest';
 
 import { SUPPLIERS } from 'tests/fixtures';
 import { where , isEqual } from 'src/operators';
+import { DEE, DUM } from 'src';
 
 describe('.where', () => {
 
@@ -28,6 +29,28 @@ describe('.where', () => {
     const res = where(input, { sid: 'S1' });
     const expected = SUPPLIERS.where({ sid: 'S1' });
     expect(isEqual(res, expected)).to.be.true;
+  })
+
+  describe('DEE and DUM', () => {
+    it('DEE.where(true) = DEE', () => {
+      const result = DEE.where(() => true);
+      expect(result.isEqual(DEE)).to.be.true;
+    })
+
+    it('DEE.where(false) = DUM', () => {
+      const result = DEE.where(() => false);
+      expect(result.isEqual(DUM)).to.be.true;
+    })
+
+    it('DUM.where(true) = DUM', () => {
+      const result = DUM.where(() => true);
+      expect(result.isEqual(DUM)).to.be.true;
+    })
+
+    it('DUM.where(false) = DUM', () => {
+      const result = DUM.where(() => false);
+      expect(result.isEqual(DUM)).to.be.true;
+    })
   })
 
 });